requirements/requirements.txt
gunicorn==22.0.0
Flask==2.3.2
Flask-Session==0.8.0
google-cloud-datacatalog==3.19.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.17.0
google-cloud-tasks==2.16.3
google-cloud-pubsub==2.21.5
google-cloud-resource-manager==1.12.3
google-cloud-logging==3.10.0
google-api-python-client==2.135.0
pytz==2024.1
pandas==2.2.2
pyarrow==16.1.0
jsonlines==4.0.0
google-auth-oauthlib==1.2.1
google-auth==2.32.0
oauth2client==4.1.3
requests==2.32.3
cachelib==0.13.0
